working building with gems on macOS

+++ Rakefile
@@ -2,7 +2,7 @@ require 'rspec/core/rake_task'
 require 'rdoc/task'
 
 # the application output name
-APPLICATION_NAME   = 'Ruþycross'
+APPLICATION_NAME    = 'Rubycross'
 
 # change this constant when you wan to change the Rubycross app version
 BUILD_VERSION      = '0.0.1'
@@ -13,6 +13,7 @@ RUBY_BUILD_VERSION = '2.2.2'
 # the main build folder
 BUILD_ROOT         = 'build/'
 BUILD_OUTPUT       = BUILD_ROOT + 'output/' + BUILD_VERSION + '/'
+BUILD_VENDOR       = BUILD_ROOT + 'vendor/'
 
 # the ouput folders for applications
 BUILD_MAC_OS       = 'macOS_'        + BUILD_VERSION + '/'
@@ -23,6 +24,8 @@ BUILD_ALL_FOLDERS  = [BUILD_MAC_OS, BUILD_LINUX_x86, BUILD_LINUX_x86_64]
 # application source folder
 SOURCE_FOLDER      = 'src/'
 OTHER_FOLDERS      = ['Users/']
+LOG_FOLDER_NAME    = 'logs/'
+GEMFILES_PATHS     = ['Gemfile', 'Gemfile.lock']
 
 # ruby file to be excecuted by the application
 EXECUTABLE_RB      = SOURCE_FOLDER + 'UI/Application.rb'
@@ -30,14 +33,31 @@ EXECUTABLE_RB      = SOURCE_FOLDER + 'UI/Application.rb'
 # application configuration
 MAC_OS_INFO_PLIST  = BUILD_ROOT + 'config/Info.plist'
 MAC_OS_APP_ICON    = BUILD_ROOT + 'config/icon.icns'
-EXECUTABLE_SCRIPT  = BUILD_ROOT + 'config/Ruþycross'
+EXECUTABLE_SCRIPT  = BUILD_ROOT + 'config/Rubycross'
 EXECUTABLE_CONTENT = 
-"#!/usr/bin/env bash
+%Q[#!/usr/bin/env bash
 
-cd `dirname $0`
-#ruby/bin/ruby #{EXECUTABLE_RB}
-/Users/Valentin/.rbenv/shims/ruby #{EXECUTABLE_RB}
-"
+# quit everything if something fail
+set -e
+
+SELFDIR="`dirname \"$0\"`"
+SELFDIR="`cd \"$SELFDIR\" && pwd`"
+
+# Tell Bundler where the Gemfile and gems are.
+export BUNDLE_GEMFILE="$SELFDIR/vendor/Gemfile"
+unset BUNDLE_IGNORE_CONFIG
+
+CURRENTDATE=`date '+%Y-%m-%d_%H_%M_%S'`
+OUTPUT_FILE="$CURRENTDATE.log"
+
+$SELFDIR/ruby/bin/ruby -rbundler/setup $SELFDIR/#{EXECUTABLE_RB} > $SELFDIR/#{LOG_FOLDER_NAME}$OUTPUT_FILE 2>&1
+
+]
+BUNDLE_CONFIG_CONTENT =
+%Q[BUNDLE_PATH: .
+BUNDLE_WITHOUT: development
+BUNDLE_DISABLE_SHARED_GEMS: '1'
+]
 
 # the Ruby binaries folder
 RUBY_BIN_ROOT         = BUILD_ROOT    + 'ruby/' + RUBY_BUILD_VERSION + '/'
@@ -88,9 +108,34 @@ task :build_prepare do
 		end
 	end
 
+	# check if vendor folder is present
+	if not File.exist?(BUILD_VENDOR) then
+		raise "Error: vendor files not presents. Install gems with `rake build_vendor`"
+	end
+
 	# create the executable script
 	File.write(EXECUTABLE_SCRIPT, EXECUTABLE_CONTENT)
 	FileUtils.chmod_R('u+x', EXECUTABLE_SCRIPT)
+
+end
+
+# generate vendor (gems)
+task :build_vendor do
+	# download all gems in the vendor folder
+	sh "env BUNDLE_IGNORE_CONFIG=1 bundle install --path #{BUILD_VENDOR} --without development"
+	
+	# add Gemfile and Gemfile.lock
+	GEMFILES_PATHS.each do |gemFile|
+		FileUtils.cp(gemFile, BUILD_VENDOR)
+	end
+
+	# remove cache files
+	sh "rm -f #{BUILD_VENDOR}*/*/cache/*"
+
+	# add .bundle folder with config inside
+	bundlePath = BUILD_VENDOR + '.bundle/'
+	FileUtils.mkdir_p(bundlePath)
+	File.write(bundlePath + 'config', BUNDLE_CONFIG_CONTENT)
 
 end
 
@@ -100,21 +145,31 @@ task :build_macOS do
 		raise "Error: #{MAC_OS_INFO_PLIST} not found!\n\n"
 	end
 
+	# main output file for macOS
 	outputFolder  = BUILD_OUTPUT + BUILD_MAC_OS + APPLICATION_NAME + '.app/'
 	outputFolder += 'Contents/'
-
 	FileUtils.mkdir_p(outputFolder)
+
+	# application informations
 	FileUtils.cp(MAC_OS_INFO_PLIST, outputFolder + 'Info.plist')
 
+	# all sources
 	sourceFolder = outputFolder + 'MacOS/'
 	FileUtils.mkdir_p(sourceFolder)
-	FileUtils.cp_r(RUBY_BIN_MAC_OS, sourceFolder + 'ruby/')
-	FileUtils.cp_r(SOURCE_FOLDER,   sourceFolder)
-	FileUtils.cp_r(OTHER_FOLDERS,   sourceFolder)
-	FileUtils.cp(EXECUTABLE_SCRIPT, sourceFolder)
 
+	FileUtils.cp_r(RUBY_BIN_MAC_OS, sourceFolder + 'ruby/') # ruby binaries
+	FileUtils.cp_r(BUILD_VENDOR,    sourceFolder) # gems
+	FileUtils.cp_r(SOURCE_FOLDER,   sourceFolder) # the souce folder
+	FileUtils.cp_r(OTHER_FOLDERS,   sourceFolder) # all other folders that are not sources
+	FileUtils.cp(EXECUTABLE_SCRIPT, sourceFolder) # the executable
+
+	FileUtils.mkdir_p(sourceFolder + LOG_FOLDER_NAME) # log folder
+
+
+	# all resources
 	resourcesFolder = outputFolder + 'Resources/'
 	FileUtils.mkdir_p(resourcesFolder)
 	FileUtils.cp(MAC_OS_APP_ICON, resourcesFolder)
+
 
 end
